+++ states_rebuilder_package/test/rm_key_test.dart
@@ -40,7 +40,11 @@ void main() {
                 rmKey = RMKey();
                 if (switcher) {
                   return StateBuilder(
-                    observeMany: [() => stringRM, () => intRM],
+                    observeMany: [
+                      () => stringRM,
+                      () => intRM,
+                      () => RM.create<String>('String')
+                    ],
                     rmKey: rmKey,
                     initState: (_, rm) {
                       rmFromInitState = rm;
@@ -62,6 +66,9 @@ void main() {
 
       await tester.pumpWidget(widget);
       rmKey.state = '';
+      expect(rmKey.get<String>().type(), '<String>');
+      expect(rmKey.get<String>(1).state, 'String');
+      expect(rmKey.get<int>(0).type(false), 'int');
       expect(rmFromInitState, equals(stringRM));
       expect(rmKey.hasObservers, isTrue);
       expect(rmKey.observers().length, 1);
@@ -268,7 +275,7 @@ void main() {
 
   testWidgets("StateBuilder should work with RMKey, Key subscription last",
       (WidgetTester tester) async {
-    RMKey<int> rmKey = RMKey();
+    RMKey<int> rmKey = RMKey(0);
 
     final widget = Builder(
       builder: (context) {
@@ -298,27 +305,27 @@ void main() {
       },
     );
     await tester.pumpWidget(widget);
-    expect(find.text('modelRM1-null'), findsOneWidget);
+    expect(find.text('modelRM1-0'), findsOneWidget);
     expect(find.text('modelRM2-0'), findsOneWidget);
-    // await tester.pump();
-    // expect(find.text('modelRM1-0'), findsOneWidget);
-    // expect(find.text('modelRM2-0'), findsOneWidget);
+    await tester.pump();
+    expect(find.text('modelRM1-0'), findsOneWidget);
+    expect(find.text('modelRM2-0'), findsOneWidget);
 
-    // rmKey.setState((_) => 1);
-    // await tester.pump();
-    // expect(find.text('modelRM1-1'), findsOneWidget);
-    // expect(find.text('modelRM2-1'), findsOneWidget);
-    // expect(rmKey.hasData, isTrue);
+    rmKey.setState((_) => 1);
+    await tester.pump();
+    expect(find.text('modelRM1-1'), findsOneWidget);
+    expect(find.text('modelRM2-1'), findsOneWidget);
+    expect(rmKey.hasData, isTrue);
 
-    // rmKey.setState((_) => 2);
-    // await tester.pump();
-    // expect(find.text('modelRM1-2'), findsOneWidget);
-    // expect(find.text('modelRM2-2'), findsOneWidget);
-    // expect(rmKey.hasData, isTrue);
-    // rmKey.refresh();
-    // await tester.pump();
-    // expect(find.text('modelRM1-0'), findsOneWidget);
-    // expect(find.text('modelRM2-0'), findsOneWidget);
+    rmKey.setState((_) => 2);
+    await tester.pump();
+    expect(find.text('modelRM1-2'), findsOneWidget);
+    expect(find.text('modelRM2-2'), findsOneWidget);
+    expect(rmKey.hasData, isTrue);
+    rmKey.refresh();
+    await tester.pump();
+    expect(find.text('modelRM1-0'), findsOneWidget);
+    expect(find.text('modelRM2-0'), findsOneWidget);
   });
 
   testWidgets(